[vertica] rename verticadb -> vertica
