+++ ddtrace/contrib/vertica/constants.py
