+++ ddtrace/contrib/vertica/patch.py
