+++ tests/contrib/vertica/utils.py
