+++ tox.ini
@@ -310,7 +310,7 @@ commands =
     sqlalchemy_contrib: nosetests {posargs} tests/contrib/sqlalchemy
     sqlite3_contrib: nosetests {posargs} tests/contrib/sqlite3
     tornado_contrib: nosetests {posargs} tests/contrib/tornado
-    vertica_contrib: pytest tests/contrib/verticadb/
+    vertica_contrib: pytest tests/contrib/vertica/
 # run subsets of the tests for particular library versions
     ddtracerun: nosetests {posargs} tests/commands/test_runner.py
     test_utils: nosetests {posargs} tests/contrib/test_utils.py
